+++ hashtable.h
@@ -91,7 +91,7 @@ void ht_set(struct ht_t *ht, const void *key, size_t key_size, const void *value
 void *ht_get(struct ht_t *ht, const void *key, size_t key_size);
 
 /* returns the first item stored with the given hash argument */
-struct ht_item_t *ht_geth(struct ht_t *ht, unsigned int hash);
+struct ht_item_t *ht_geth(struct ht_t *ht, size_t hash);
 
 /* removes and frees the item the hashtable */
 void ht_rm(struct ht_t *ht, const void *key, size_t key_size);
